requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
aiosqlite==0.22.1
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
auth0-ai==1.0.2
auth0-ai-llamaindex==1.0.1
auth0-python==4.13.0
banks==2.4.1
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
cryptography==46.0.6
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.3.1
dirtyjson==1.0.8
distro==1.9.0
filelock==3.25.2
filetype==1.2.0
frozenlist==1.8.0
fsspec==2026.3.0
greenlet==3.3.2
griffe==2.0.2
griffecli==2.0.2
griffelib==2.0.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
idna==3.11
importlib_metadata==8.7.1
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
llama-index==0.14.16
llama-index-agent-openai==0.4.12
llama-index-cli==0.5.5
llama-index-core==0.14.20
llama-index-embeddings-openai==0.5.2
llama-index-indices-managed-llama-cloud==0.11.1
llama-index-instrumentation==0.5.0
llama-index-llms-groq==0.5.0
llama-index-llms-openai==0.6.26
llama-index-llms-openai-like==0.5.3
llama-index-program-openai==0.3.2
llama-index-readers-file==0.5.6
llama-index-readers-llama-parse==0.6.1
llama-index-workflows==2.17.1
llama-parse==0.5.20
llama_cloud==1.6.0
MarkupSafe==3.0.3
marshmallow==3.26.2
multidict==6.7.1
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.6.1
nltk==3.9.4
numpy==2.4.4
openai==1.109.1
openfga-sdk==0.9.9
opentelemetry-api==1.40.0
packaging==26.0
pandas==2.3.3
pillow==12.2.0
platformdirs==4.9.4
propcache==0.4.1
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.12.1
pypdf==6.9.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
regex==2026.4.4
requests==2.33.1
safetensors==0.7.0
setuptools==82.0.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.49
striprtf==0.0.26
tenacity==9.1.4
tiktoken==0.12.0
tinytag==2.2.1
tokenizers==0.22.2
tqdm==4.67.3
transformers==4.57.6
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
wrapt==2.1.2
yarl==1.23.0
zipp==3.23.0
